fix: only chown backup root on init
In the event backup user is root, any permissions can be set on the
parent folder. If a client rsyncs the root module folder, the owner of
this folder could change, creating a competition for ownership.

Setting this once ensures the user can write to it, but subsequent
permissions can be governed by client backup routines.